Refactored packages and file structure
