+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.NasaDailyImage"
         tools:targetApi="31">
         <activity
-            android:name=".About"
+            android:name=".activity.About"
             android:exported="false"
             android:launchMode="singleTask"
             android:screenOrientation = "portrait">
@@ -24,7 +24,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".ListImage"
+            android:name=".activity.ListImage"
             android:exported="false"
             android:launchMode="singleTask"
             android:screenOrientation = "portrait">
@@ -33,7 +33,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".SearchImage"
+            android:name=".activity.SearchImage"
             android:exported="false"
             android:launchMode="singleTask"
             android:screenOrientation = "portrait">
@@ -42,7 +42,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".RandomImage"
+            android:name=".activity.RandomImage"
             android:exported="false"
             android:launchMode="singleTask"
             android:screenOrientation = "portrait">
@@ -51,7 +51,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true"
             android:launchMode="singleTask"
             android:screenOrientation = "portrait">
