--- app/src/main/java/com/jr_dev/nasadailyimage/WebQuery.java
@@ -1,4 +1,0 @@
-package com.jr_dev.nasadailyimage;
-
-public class WebQuery {
-}
