+++ app/src/main/java/com/jr_dev/nasadailyimage/activity/SearchImage.java
@@ -1,4 +1,4 @@
-package com.jr_dev.nasadailyimage;
+package com.jr_dev.nasadailyimage.activity;
 
 
 import android.content.Intent;
@@ -24,6 +24,9 @@ import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 
 import com.google.android.material.navigation.NavigationView;
+import com.jr_dev.nasadailyimage.data.ImageDAO;
+import com.jr_dev.nasadailyimage.R;
+import com.jr_dev.nasadailyimage.component.SearchFragment;
 
 import org.json.JSONException;
 import org.json.JSONObject;
