+++ app/src/main/java/com/jr_dev/nasadailyimage/data/ImageDB.java
@@ -1,4 +1,4 @@
-package com.jr_dev.nasadailyimage;
+package com.jr_dev.nasadailyimage.data;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
