+++ app/src/main/java/com/jr_dev/nasadailyimage/data/WebQuery.java
@@ -1,0 +1,4 @@
+package com.jr_dev.nasadailyimage.data;
+
+public class WebQuery {
+}
